Add a dedicated controller for individual CSV importers and simplify form class.
